final fixes with types
