+++ src/components/TodoInfo/TodoInfo.tsx
@@ -1,17 +1,17 @@
 import cn from 'classnames';
 import { UserInfo } from '../UserInfo';
 import { FC } from 'react';
-import { Item } from '../TodoList/TodoList';
+import { TodoWithUser } from '../../types/Types';
 
-export const TodoInfo: FC<{ todo: Item }> = ({ todo }) => {
+export const TodoInfo: FC<{ todo: TodoWithUser }> = ({ todo }) => {
   return (
     <article
-      data-id={todo.toDoItem.id}
+      data-id={todo.id}
       className={cn('TodoInfo', {
-        'TodoInfo--completed': todo.toDoItem.completed,
+        'TodoInfo--completed': todo.completed,
       })}
     >
-      <h2 className="TodoInfo__title">{todo.toDoItem.title}</h2>
+      <h2 className="TodoInfo__title">{todo.title}</h2>
       <UserInfo user={todo.user} />
     </article>
   );
