+++ src/components/TodoList/TodoList.tsx
@@ -1,21 +1,16 @@
 import { FC } from 'react';
-import { ToDoItem, User } from '../../App';
+import { TodoWithUser } from '../../types/Types';
 import { TodoInfo } from '../TodoInfo';
 
-export interface Item {
-  toDoItem: ToDoItem;
-  user: User;
-}
-
 interface Props {
-  todos: Item[];
+  todos: TodoWithUser[];
 }
 
 export const TodoList: FC<Props> = ({ todos }) => {
   return (
     <section className="TodoList">
       {todos.map(item => (
-        <TodoInfo todo={item} key={item.toDoItem.id} />
+        <TodoInfo todo={item} key={item.id} />
       ))}
     </section>
   );
