+++ src/types/Types.ts
@@ -1,0 +1,17 @@
+export interface ToDoItem {
+  id: number;
+  title: string;
+  completed: boolean;
+  userId: number;
+}
+
+export interface User {
+  id: number;
+  name: string;
+  username: string;
+  email: string;
+}
+
+export interface TodoWithUser extends ToDoItem {
+  user: User;
+}
